added gates for comments and posts for more protection

+++ app/Http/Controllers/PostCommentsController.php
@@ -14,15 +14,29 @@ class PostCommentsController extends Controller
         $data = $request->validated();
         if($request->hasFile('picture'))
         {
-            $data['picture'] = $request->file('picture')->store('pictures/posts', 'public');
+            $data['picture'] = $request->file('picture')->store('pictures/comments', 'public');
         }
         Comment::create(array_merge($data, ['user_id' => Auth::id(), 'post_id' => $request['post_id']]));
         return back()->with('message', 'Comment successfully created');
     }
 
+    public function update(Comment $comment, CommentRequest $request)
+    {
+        $this->authorize('update-comment',  $comment);
+        $data = $request->validated();
+        if($request->hasFile('picture'))
+        {
+            $file = $request->file('picture');
+            $path = $file->store('pictures/comments', 'public');
+            $data['picture'] = $path;
+        }
+        $comment->update($data);
+        return back()->with('message', "Comment successfully updated");
+    }
+
     public function destroy(Post $post, Comment $comment)
     {
-        //$this->authorize('delete', $comment);
+        $this->authorize('delete-comment', $comment);
         $comment->delete();
         return back()->with('message', 'Comment deleted');
     }
